+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- [#64] Add retry logic for carpenter build.
 - [#61] Add full API documentation
 - [#60] Add onboarding guide for setting up a new project with an existing warehouse.ai service
 - [#56] Extract diagram tool for other Warehouse.ai modules.
@@ -29,3 +30,4 @@
 [#59]: https://github.com/godaddy/warehouse.ai/pull/59
 [#60]: https://github.com/godaddy/warehouse.ai/pull/60
 [#61]: https://github.com/godaddy/warehouse.ai/pull/61
+[#64]: https://github.com/godaddy/warehouse.ai/pull/64